chore: share test generators through commons
