+++ packages/commons/src/events/index.ts
@@ -36,6 +36,7 @@ export * from './ActionType'
 export * from './FieldTypeMapping'
 export * from './Conditional'
 export * from './AdvancedSearchConfig'
+export * from './test.utils'
 // In order to infer types, we need to export the following types along with events
 export * from '../conditionals/conditionals'
 export * from '../conditionals/validate'
