+++ packages/commons/src/events/test.utils.ts
@@ -1,0 +1,147 @@
+import { tennisClubMembershipEvent } from '../fixtures'
+import { getUUID } from '../uuid'
+import {
+  DeclareActionInput,
+  RegisterActionInput,
+  RequestCorrectionActionInput,
+  ValidateActionInput
+} from './ActionInput'
+import { ActionType } from './ActionType'
+import { EventConfig } from './EventConfig'
+import { EventInput } from './EventInput'
+import { mapFieldTypeToMockValue } from './FieldTypeMapping'
+import { findActiveActionFields, stripHiddenOrDisabledFields } from './utils'
+
+export function generateActionInput(
+  configuration: EventConfig,
+  action: ActionType
+) {
+  const fields = findActiveActionFields(configuration, action) ?? []
+
+  const data = fields.reduce(
+    (acc, field, i) => ({
+      ...acc,
+      [field.id]: mapFieldTypeToMockValue(field, i)
+    }),
+    {}
+  )
+
+  // Strip away hidden or disabled fields from mock action data
+  // If this is not done, the mock data might contain hidden or disabled fields, which will cause validation errors
+  return stripHiddenOrDisabledFields(action, configuration, data)
+}
+
+export const eventPayloadGenerator = {
+  create: (input: Partial<EventInput> = {}) => ({
+    transactionId: input.transactionId ?? getUUID(),
+    type: input.type ?? 'TENNIS_CLUB_MEMBERSHIP'
+  }),
+  patch: (id: string, input: Partial<EventInput> = {}) => ({
+    transactionId: input.transactionId ?? getUUID(),
+    type: input.type ?? 'TENNIS_CLUB_MEMBERSHIP',
+    id
+  }),
+  actions: {
+    declare: (
+      eventId: string,
+      input: Partial<Pick<DeclareActionInput, 'transactionId' | 'data'>> = {}
+    ) => ({
+      type: ActionType.DECLARE,
+      transactionId: input.transactionId ?? getUUID(),
+      data:
+        input.data ??
+        generateActionInput(tennisClubMembershipEvent, ActionType.DECLARE),
+      eventId
+    }),
+    validate: (
+      eventId: string,
+      input: Partial<Pick<ValidateActionInput, 'transactionId' | 'data'>> = {}
+    ) => ({
+      type: ActionType.VALIDATE,
+      transactionId: input.transactionId ?? getUUID(),
+      data: input.data ?? {},
+      duplicates: [],
+      eventId
+    }),
+    register: (
+      eventId: string,
+      input: Partial<Pick<RegisterActionInput, 'transactionId' | 'data'>> = {}
+    ) => ({
+      type: ActionType.REGISTER,
+      transactionId: input.transactionId ?? getUUID(),
+      data:
+        input.data ??
+        generateActionInput(tennisClubMembershipEvent, ActionType.REGISTER),
+      eventId
+    }),
+    printCertificate: (
+      eventId: string,
+      input: Partial<Pick<RegisterActionInput, 'transactionId' | 'data'>> = {}
+    ) => ({
+      type: ActionType.PRINT_CERTIFICATE,
+      transactionId: input.transactionId ?? getUUID(),
+      data:
+        input.data ??
+        generateActionInput(
+          tennisClubMembershipEvent,
+          ActionType.PRINT_CERTIFICATE
+        ),
+      eventId
+    }),
+    correction: {
+      request: (
+        eventId: string,
+        input: Partial<
+          Pick<RequestCorrectionActionInput, 'transactionId' | 'data'>
+        > = {}
+      ) => ({
+        type: ActionType.REQUEST_CORRECTION,
+        transactionId: input.transactionId ?? getUUID(),
+        data:
+          input.data ??
+          generateActionInput(
+            tennisClubMembershipEvent,
+            ActionType.REQUEST_CORRECTION
+          ),
+        metadata: {},
+        eventId
+      }),
+      approve: (
+        eventId: string,
+        requestId: string,
+        input: Partial<
+          Pick<RequestCorrectionActionInput, 'transactionId' | 'data'>
+        > = {}
+      ) => ({
+        type: ActionType.APPROVE_CORRECTION,
+        transactionId: input.transactionId ?? getUUID(),
+        data:
+          input.data ??
+          generateActionInput(
+            tennisClubMembershipEvent,
+            ActionType.APPROVE_CORRECTION
+          ),
+        eventId,
+        requestId
+      }),
+      reject: (
+        eventId: string,
+        requestId: string,
+        input: Partial<
+          Pick<RequestCorrectionActionInput, 'transactionId' | 'data'>
+        > = {}
+      ) => ({
+        type: ActionType.REJECT_CORRECTION,
+        transactionId: input.transactionId ?? getUUID(),
+        data:
+          input.data ??
+          generateActionInput(
+            tennisClubMembershipEvent,
+            ActionType.REJECT_CORRECTION
+          ),
+        eventId,
+        requestId
+      })
+    }
+  }
+}
