+++ packages/events/src/tests/generators.ts
@@ -20,30 +20,11 @@ import {
   mapFieldTypeToMockValue,
   EventConfig,
   findActiveActionFields,
-  stripHiddenOrDisabledFields
+  stripHiddenOrDisabledFields,
+  eventPayloadGenerator
 } from '@opencrvs/commons'
 import { Location } from '@events/service/locations/locations'
 import { Db } from 'mongodb'
-import { tennisClubMembershipEvent } from '@opencrvs/commons/fixtures'
-
-export function generateActionInput(
-  configuration: EventConfig,
-  action: ActionType
-) {
-  const fields = findActiveActionFields(configuration, action) ?? []
-
-  const data = fields.reduce(
-    (acc, field, i) => ({
-      ...acc,
-      [field.id]: mapFieldTypeToMockValue(field, i)
-    }),
-    {}
-  )
-
-  // Strip away hidden or disabled fields from mock action data
-  // If this is not done, the mock data might contain hidden or disabled fields, which will cause validation errors
-  return stripHiddenOrDisabledFields(action, configuration, data)
-}
 
 interface Name {
   use: string
@@ -68,121 +49,6 @@ interface CreateUser {
  * @returns a payload generator for creating events and actions with sensible defaults.
  */
 export function payloadGenerator() {
-  const event = {
-    create: (input: Partial<EventInput> = {}) => ({
-      transactionId: input.transactionId ?? getUUID(),
-      type: input.type ?? 'TENNIS_CLUB_MEMBERSHIP'
-    }),
-    patch: (id: string, input: Partial<EventInput> = {}) => ({
-      transactionId: input.transactionId ?? getUUID(),
-      type: input.type ?? 'TENNIS_CLUB_MEMBERSHIP',
-      id
-    }),
-    actions: {
-      declare: (
-        eventId: string,
-        input: Partial<Pick<DeclareActionInput, 'transactionId' | 'data'>> = {}
-      ) => ({
-        type: ActionType.DECLARE,
-        transactionId: input.transactionId ?? getUUID(),
-        data:
-          input.data ??
-          generateActionInput(tennisClubMembershipEvent, ActionType.DECLARE),
-        eventId
-      }),
-      validate: (
-        eventId: string,
-        input: Partial<Pick<ValidateActionInput, 'transactionId' | 'data'>> = {}
-      ) => ({
-        type: ActionType.VALIDATE,
-        transactionId: input.transactionId ?? getUUID(),
-        data: input.data ?? {},
-        duplicates: [],
-        eventId
-      }),
-      register: (
-        eventId: string,
-        input: Partial<Pick<RegisterActionInput, 'transactionId' | 'data'>> = {}
-      ) => ({
-        type: ActionType.REGISTER,
-        transactionId: input.transactionId ?? getUUID(),
-        data:
-          input.data ??
-          generateActionInput(tennisClubMembershipEvent, ActionType.REGISTER),
-        eventId
-      }),
-      printCertificate: (
-        eventId: string,
-        input: Partial<Pick<RegisterActionInput, 'transactionId' | 'data'>> = {}
-      ) => ({
-        type: ActionType.PRINT_CERTIFICATE,
-        transactionId: input.transactionId ?? getUUID(),
-        data:
-          input.data ??
-          generateActionInput(
-            tennisClubMembershipEvent,
-            ActionType.PRINT_CERTIFICATE
-          ),
-        eventId
-      }),
-      correction: {
-        request: (
-          eventId: string,
-          input: Partial<
-            Pick<RequestCorrectionActionInput, 'transactionId' | 'data'>
-          > = {}
-        ) => ({
-          type: ActionType.REQUEST_CORRECTION,
-          transactionId: input.transactionId ?? getUUID(),
-          data:
-            input.data ??
-            generateActionInput(
-              tennisClubMembershipEvent,
-              ActionType.REQUEST_CORRECTION
-            ),
-          metadata: {},
-          eventId
-        }),
-        approve: (
-          eventId: string,
-          requestId: string,
-          input: Partial<
-            Pick<RequestCorrectionActionInput, 'transactionId' | 'data'>
-          > = {}
-        ) => ({
-          type: ActionType.APPROVE_CORRECTION,
-          transactionId: input.transactionId ?? getUUID(),
-          data:
-            input.data ??
-            generateActionInput(
-              tennisClubMembershipEvent,
-              ActionType.APPROVE_CORRECTION
-            ),
-          eventId,
-          requestId
-        }),
-        reject: (
-          eventId: string,
-          requestId: string,
-          input: Partial<
-            Pick<RequestCorrectionActionInput, 'transactionId' | 'data'>
-          > = {}
-        ) => ({
-          type: ActionType.REJECT_CORRECTION,
-          transactionId: input.transactionId ?? getUUID(),
-          data:
-            input.data ??
-            generateActionInput(
-              tennisClubMembershipEvent,
-              ActionType.REJECT_CORRECTION
-            ),
-          eventId,
-          requestId
-        })
-      }
-    }
-  }
-
   const user = {
     create: (input: CreateUser) => ({
       role: input.role ?? 'REGISTRATION_AGENT',
@@ -212,7 +78,7 @@ export function payloadGenerator() {
     }
   }
 
-  return { event, locations, user }
+  return { event: eventPayloadGenerator, locations, user }
 }
 
 /**
